feat(login): add login page with form and input validation

+++ server/src/model/ createTable.ts
@@ -13,7 +13,7 @@ const createCoffeeTable = async () => {
         id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
         name TEXT NOT NULL,
         description TEXT NOT NULL,
-        price INT NOT NULL,
+        price NUMERIC(10,2) NOT NULL,
         imagePath TEXT NOT NULL,
         isFeatured BOOLEAN NOT NULL)`);
   } catch (err) {
@@ -52,5 +52,3 @@ const seedAdminUser = async () => {
     throw err;
   }
 };
-
-//TODO change int to numeric
